requirements.txt
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bs4==0.0.1
capstone==4.0.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
decorator==4.4.2
discord-webhook==0.9.0
discord.py==1.4.1
enum34==1.1.10
idna==2.10
intervaltree==3.1.0
ipaddress==1.0.23
ipython==5.10.0
ipython-genutils==0.2.0
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.4
paramiko==2.7.2
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==1.0.18
psutil==5.7.2
ptyprocess==0.6.0
pwntools==4.2.1
pycparser==2.20
pyelftools==0.26
Pygments==2.5.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.24.0
ROPGadget==6.3
scandir==1.10.0
simplegeneric==0.8.1
six==1.15.0
sortedcontainers==2.2.2
soupsieve==1.9.6
traitlets==4.3.3
unicorn==1.0.2rc3
urllib3==1.25.10
wcwidth==0.2.5
backports.functools-lru-cache==1.6.1
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bs4==0.0.1
capstone==4.0.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
decorator==4.4.2
discord-webhook==0.9.0
discord.py==1.4.1
enum34==1.1.10
idna==2.10
intervaltree==3.1.0
ipaddress==1.0.23
ipython==5.10.0
ipython-genutils==0.2.0
lxml==4.5.2
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.4
paramiko==2.7.2
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==1.0.18
psutil==5.7.2
ptyprocess==0.6.0
pwntools==4.2.1
pycparser==2.20
pyelftools==0.26
Pygments==2.5.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.24.0
ROPGadget==6.3
scandir==1.10.0
simplegeneric==0.8.1
six==1.15.0
sortedcontainers==2.2.2
soupsieve==1.9.6
traitlets==4.3.3
unicorn==1.0.2rc3
urllib3==1.25.10
wcwidth==0.2.5